+++ src/util/init.lua
@@ -1,2 +1,0 @@
--- 初始化顏色
-require 'color':Init()