+++ src/util/random_number_generator.lua
@@ -3,22 +3,33 @@
 -- 一開始創建要給name_、probability_(小於等於100的整數)。
 
 local math = math
+local require = require
 math.randomseed(tostring(os.time()):reverse():sub(1, 6))
 
-local RNG = require 'class'("RNG")
+local RNG = require 'util.class'("RNG")
 
 -- assert
 local GenerateDeck, GenerateValidCards
 
-function RNG:_new(cards_label, probability)
-    GenerateDeck(self, probability)
-    GenerateValidCards(self)
+function RNG:_new(probability, cards_label)
+    local instance = {
+        _used_card_count_  = 0,
+        _valid_card_count_ = 0,
+        _total_card_count_ = 0
+    }
 
-    RNG:setInstance(cards_label, self)
+    GenerateDeck(instance, probability)
+    GenerateValidCards(instance)
+
+    if cards_label then
+        RNG:setInstance(cards_label, instance)
+    end
+
+    return instance
 end
 
 GenerateDeck = function(self, probability)
-    local gcf = require 'greatest_common_factor'
+    local gcf = require 'util.greatest_common_factor'
     
     -- 計算牌數
     local gcd = gcf(probability, 100)
@@ -40,26 +51,25 @@ end
 -- 抽牌
 function RNG:draw()
     if self._used_card_count_ >= self._total_card_count_ then
-        self:Reset()
+        self:reset()
     end
 
     self._used_card_count_ = self._used_card_count_ + 1
     return self[self._used_card_count_]
 end
 
-function RNG:Reset()
+function RNG:reset()
     self._used_card_count_ = 0
     GenerateValidCards(self)
 end
 
 -- 創建有效牌和無效牌
 GenerateValidCards = function(self)
-    local rand = require 'math_lib'.Random
+    local rand = require 'util.math_lib'.rand
 
     local current_valid_card_count = 0
     for i = 1, self._total_card_count_ do
-        --rand()只會生成 0, 1 兩個數，所以可以分別有效、無效
-        if (rand() == 1) and (current_valid_card_count < self._valid_card_count_) then
+        if (rand() > 0.5) and (current_valid_card_count < self._valid_card_count_) then
             self[i] = true
             current_valid_card_count = current_valid_card_count + 1
         else
@@ -79,6 +89,10 @@ GenerateValidCards = function(self)
             demands = demands - 1
         end
     end
+end
+
+function RNG:count()
+    return self._total_card_count_
 end
 
 return RNG